Added a platform the player can interact with
The player can jump on the platform, and falls if it steps off.

Programmed this along with Curtis (9) and Eli (7). They were flanking me
while sitting on the couch, and it was quite possibly the most fun I've
ever had programming.

They both were totally engaged in it, contributing ideas, laughing at
the various glitches along the way (when the "guy" wouldn't fall, kept
jumping, fell through the floor, etc) and overall being amazed at how
much work it took to simply get our box-shaped character to do what we
want on the screen.